Support binary bodies in responses
`aioresponses` type annotations do not allow passing instances of
`bytes` as the `body` parameter even though the library does correctly
support such bodies:

    def _build_response(self, url: 'Union[URL, str]',
                        method: str = hdrs.METH_GET,
                        request_headers: Dict = None,
                        status: int = 200,
                        body: str = '',
    (...)
                        reason: Optional[str] = None) -> ClientResponse:
        if response_class is None:
            response_class = ClientResponse
        if payload is not None:
            body = json.dumps(payload)
        if not isinstance(body, bytes):
            body = str.encode(body)

To allow passing binary bodies, we change type annotations for the
`body` argument from `str` to `Union[str, bytes]`. No more changes are
needed because `bytes` are already handled correctly by the code in
`_build_response` quoted above. To prove that it works, we add a couple
tests that set the body to a byte sequence which is not valid UTF-8 and
any attempts to encode it would break it.