Drop Python 3.6 support
We cannot run tests with Python 3.10 using Pytest 6.x due to [1]. This
is fixed in Pytest 7.x but Pytest 7.x no longer supports Python 3.6
which aioresponses tries to support. However because Python 3.6 is
already past the EOL date, the simplest solution is to no longer test
with Python 3.6 and to stop declaring support for that version.

[1] https://github.com/pytest-dev/pytest/issues/8546